Moved from framework to core
