+++ AFB.sln
@@ -5,11 +5,11 @@ VisualStudioVersion = 16.0.31005.135
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AFB.Console", "AFB.Console\AFB.Console.csproj", "{8D8C0C44-82E7-4176-8B4E-A383BBB1B015}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AFB.Desktop", "AFB.Desktop\AFB.Desktop.csproj", "{EE178E79-40FF-4D0C-8E56-C7F94CA20A14}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleApp1", "ConsoleApp1\ConsoleApp1.csproj", "{24A877BB-5C5F-4810-A97B-35F5FC3F7EC7}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConsoleApp1", "ConsoleApp1\ConsoleApp1.csproj", "{24A877BB-5C5F-4810-A97B-35F5FC3F7EC7}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AFB.Logic", "AFB.Logic\AFB.Logic.csproj", "{A624B0C3-091B-45D1-97AC-61AB1B814B11}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AFB.Logic", "AFB.Logic\AFB.Logic.csproj", "{A624B0C3-091B-45D1-97AC-61AB1B814B11}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AFB.Desktop", "AFB.Desktop\AFB.Desktop.csproj", "{5FF07EDC-0649-4C7C-A165-253281739096}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,10 +21,6 @@ Global
 		{8D8C0C44-82E7-4176-8B4E-A383BBB1B015}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8D8C0C44-82E7-4176-8B4E-A383BBB1B015}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8D8C0C44-82E7-4176-8B4E-A383BBB1B015}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EE178E79-40FF-4D0C-8E56-C7F94CA20A14}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EE178E79-40FF-4D0C-8E56-C7F94CA20A14}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EE178E79-40FF-4D0C-8E56-C7F94CA20A14}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EE178E79-40FF-4D0C-8E56-C7F94CA20A14}.Release|Any CPU.Build.0 = Release|Any CPU
 		{24A877BB-5C5F-4810-A97B-35F5FC3F7EC7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{24A877BB-5C5F-4810-A97B-35F5FC3F7EC7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{24A877BB-5C5F-4810-A97B-35F5FC3F7EC7}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -33,6 +29,10 @@ Global
 		{A624B0C3-091B-45D1-97AC-61AB1B814B11}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A624B0C3-091B-45D1-97AC-61AB1B814B11}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A624B0C3-091B-45D1-97AC-61AB1B814B11}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5FF07EDC-0649-4C7C-A165-253281739096}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5FF07EDC-0649-4C7C-A165-253281739096}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5FF07EDC-0649-4C7C-A165-253281739096}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5FF07EDC-0649-4C7C-A165-253281739096}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
