+++ WpfApp1/App.xaml.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows;
 
-namespace AFB.Desktop
+namespace WpfApp1
 {
     /// <summary>
     /// Interaction logic for App.xaml
